db: compute file count and size metrics incrementally
Previously, level sizes were recomputed from scratch with each new
version. This showed up in CPU profiles of stores with large numbers of
sstables (#718). This change updates logAndApply callers to populate the
passed LevelMetrics' `NumFiles` and `Size` fields with the incremental
diff of the version edit.

+++ metrics_test.go
@@ -47,7 +47,7 @@ func TestMetricsFormat(t *testing.T) {
 		base := uint64((i + 1) * 100)
 		l.Sublevels = int32(i + 1)
 		l.NumFiles = int64(base) + 1
-		l.Size = base + 2
+		l.Size = int64(base) + 2
 		l.Score = float64(base) + 3
 		l.BytesIn = base + 4
 		l.BytesIngested = base + 4
